review-agent/pyproject.toml
[project]
name = "review-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.0",
    "google-cloud-aiplatform>=1.125.0",
    "google-genai>=1.48.0",
    "google-generativeai>=0.8.5",
    "langchain>=1.0.3",
    "langchain-core>=1.0.3",
    "langgraph>=1.0.2",
    "ollama>=0.6.0",
    "openai>=2.7.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "together>=1.5.30",
    "uvicorn>=0.38.0",
]
